pendiente UPDATE

+++ Spring Boot Project/src/main/java/com/ejemplo/tiendaalamano/data/entities/MarcaEntity.java
@@ -17,7 +17,7 @@ public class MarcaEntity implements Serializable{
     @Id
     @Column(name="id_marca")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idMarca;
+    private Integer idMarca;
 
     @Column(nullable = true, length = 50)
     private String nombre;
@@ -29,11 +29,11 @@ public class MarcaEntity implements Serializable{
     //private List<ProductoEntity> productoModelList = new ArrayList<>();
     
 
-    public Long getIdMarca() {
+    public Integer getIdMarca() {
         return idMarca;
     }
 
-    public void setIdMarca(Long idMarca) {
+    public void setIdMarca(Integer idMarca) {
         this.idMarca = idMarca;
     }
